requirements.txt
streamlit
transformers
pinecone-client
sentence_transformers
openai
